Scoring system
This is the scoring system library. I have only ran it once. I am currently making the minheap to sort these, as mentioned in the instructions. I believe there is room for the player IDs to be written in the files so that leaderboard can be implemented, but i have not done that because player profiles have not been made yet.